Move addToProject to PathUtils
