+++ src/Mod/Path/PathScripts/PathProfile.py
@@ -257,7 +257,7 @@ class CommandPathProfile:
         obj.Active = True
         obj.ViewObject.ShowFirstRapid = False
 
-        project = PathProject.CommandProject.addToProject(obj)
+        project = PathUtils.addToProject(obj)
 
         tl = PathUtils.changeTool(obj,project)
         if tl:
